Update 9/2
- Added LWA+Mirror to the Inverted Blacksmith check.

Unsure what I'm going to do for v31. All of the new settings (Restricted vs. Advanced item placement, MCS Shuffle, etc.) really changes the way that the algorithm handles things and is likely going to have a lot of minor effects on the logic that will all have to be accounted for with all sorts of different knobs and levers that the tracker's going to have to implement. If I do end up putting a v31 tracker together it'll probably take at least a couple weeks after the version is released. If not, it's been a good run.